# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Finventory API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  company?: string;
  subject?: string;
  message: string;
}

export type ContactStatus = typeof ContactStatus[keyof typeof ContactStatus];


export const ContactStatus = {
  new: 'new',
  read: 'read',
  replied: 'replied',
  closed: 'closed',
} as const;

export interface Contact {
  id: string;
  name: string;
  email: string;
  phone?: string;
  company?: string;
  subject?: string;
  message: string;
  status: ContactStatus;
  createdAt: string;
}

export interface ContactListResponse {
  data: Contact[];
  total: number;
  page: number;
  totalPages: number;
}

export interface DemoBookingInput {
  fullName: string;
  businessName: string;
  industry: string;
  phone: string;
  email: string;
  companySize: string;
  preferredDate: string;
  preferredTime: string;
  message?: string;
}

export type DemoBookingStatus = typeof DemoBookingStatus[keyof typeof DemoBookingStatus];


export const DemoBookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface DemoBooking {
  id: string;
  fullName: string;
  businessName: string;
  industry: string;
  phone: string;
  email: string;
  companySize: string;
  preferredDate: string;
  preferredTime: string;
  message?: string;
  status: DemoBookingStatus;
  createdAt: string;
}

export interface DemoBookingListResponse {
  data: DemoBooking[];
  total: number;
  page: number;
  totalPages: number;
}

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  excerpt: string;
  content?: string;
  category: string;
  tags?: string[];
  coverImage?: string;
  author: string;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostInput {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  tags?: string[];
  coverImage?: string;
  author: string;
  published?: boolean;
}

export interface BlogListResponse {
  data: BlogPost[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Faq {
  id: string;
  question: string;
  answer: string;
  category: string;
  order?: number;
}

export interface Testimonial {
  id: string;
  name: string;
  role: string;
  company: string;
  avatar?: string;
  content: string;
  rating: number;
  industry?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type LoginResponseUser = {
  id: string;
  email: string;
  name: string;
};

export interface LoginResponse {
  token: string;
  user: LoginResponseUser;
}

export interface StatusUpdate {
  status: string;
}

export interface AdminStats {
  totalContacts: number;
  newContacts: number;
  totalDemoBookings: number;
  pendingDemoBookings: number;
  totalBlogPosts: number;
  publishedBlogPosts: number;
}

export type ListBlogPostsParams = {
page?: number;
limit?: number;
category?: string;
};

export type ListFaqsParams = {
category?: string;
};

export type ListAdminContactsParams = {
page?: number;
limit?: number;
status?: ListAdminContactsStatus;
};

export type ListAdminContactsStatus = typeof ListAdminContactsStatus[keyof typeof ListAdminContactsStatus];


export const ListAdminContactsStatus = {
  new: 'new',
  read: 'read',
  replied: 'replied',
  closed: 'closed',
} as const;

export type ListAdminDemoBookingsParams = {
page?: number;
limit?: number;
status?: ListAdminDemoBookingsStatus;
};

export type ListAdminDemoBookingsStatus = typeof ListAdminDemoBookingsStatus[keyof typeof ListAdminDemoBookingsStatus];


export const ListAdminDemoBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;