# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Finventory API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact enquiries
  - name: demo
    description: Demo bookings
  - name: blog
    description: Blog posts
  - name: faq
    description: FAQs
  - name: testimonial
    description: Testimonials
  - name: auth
    description: Authentication
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contacts:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /demo-bookings:
    post:
      operationId: submitDemoBooking
      tags: [demo]
      summary: Submit a demo booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoBookingInput"
      responses:
        "201":
          description: Demo booking submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 9
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"

  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /faqs:
    get:
      operationId: listFaqs
      tags: [faq]
      summary: List all FAQs
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: List of FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"

  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonial]
      summary: List testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/contacts:
    get:
      operationId: listAdminContacts
      tags: [admin]
      summary: List all contact enquiries
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum: [new, read, replied, closed]
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"

  /admin/contacts/{id}:
    patch:
      operationId: updateContactStatus
      tags: [admin]
      summary: Update contact enquiry status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/demo-bookings:
    get:
      operationId: listAdminDemoBookings
      tags: [admin]
      summary: List all demo bookings
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
      responses:
        "200":
          description: List of demo bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoBookingListResponse"

  /admin/demo-bookings/{id}:
    patch:
      operationId: updateDemoBookingStatus
      tags: [admin]
      summary: Update demo booking status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/blog:
    get:
      operationId: listAdminBlogPosts
      tags: [admin]
      summary: List all blog posts (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
    post:
      operationId: createBlogPost
      tags: [admin]
      summary: Create a blog post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  /admin/blog/{id}:
    put:
      operationId: updateBlogPost
      tags: [admin]
      summary: Update a blog post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [admin]
      summary: Delete a blog post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        company:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - message

    Contact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [new, read, replied, closed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - message
        - status
        - createdAt

    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages

    DemoBookingInput:
      type: object
      properties:
        fullName:
          type: string
        businessName:
          type: string
        industry:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        companySize:
          type: string
        preferredDate:
          type: string
        preferredTime:
          type: string
        message:
          type: string
      required:
        - fullName
        - businessName
        - industry
        - phone
        - email
        - companySize
        - preferredDate
        - preferredTime

    DemoBooking:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        businessName:
          type: string
        industry:
          type: string
        phone:
          type: string
        email:
          type: string
        companySize:
          type: string
        preferredDate:
          type: string
        preferredTime:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - businessName
        - industry
        - phone
        - email
        - companySize
        - preferredDate
        - preferredTime
        - status
        - createdAt

    DemoBookingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DemoBooking"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages

    BlogPost:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        coverImage:
          type: string
        author:
          type: string
        published:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - excerpt
        - category
        - author
        - published
        - createdAt
        - updatedAt

    BlogPostInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        coverImage:
          type: string
        author:
          type: string
        published:
          type: boolean
      required:
        - title
        - slug
        - excerpt
        - content
        - category
        - author

    BlogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages

    Faq:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
        category:
          type: string
        order:
          type: integer
      required:
        - id
        - question
        - answer
        - category

    Testimonial:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        company:
          type: string
        avatar:
          type: string
        content:
          type: string
        rating:
          type: integer
        industry:
          type: string
      required:
        - id
        - name
        - role
        - company
        - content
        - rating

    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            name:
              type: string
          required:
            - id
            - email
            - name
      required:
        - token
        - user

    StatusUpdate:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AdminStats:
      type: object
      properties:
        totalContacts:
          type: integer
        newContacts:
          type: integer
        totalDemoBookings:
          type: integer
        pendingDemoBookings:
          type: integer
        totalBlogPosts:
          type: integer
        publishedBlogPosts:
          type: integer
      required:
        - totalContacts
        - newContacts
        - totalDemoBookings
        - pendingDemoBookings
        - totalBlogPosts
        - publishedBlogPosts