# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Finventory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  excerpt: string;
  content?: string;
  category: string;
  tags?: string[];
  coverImage?: string;
  author: string;
  published: boolean;
  createdAt: Date;
  updatedAt: Date;
}