# lib/api-zod/src/generated/types/contactListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Finventory API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from './contact';

export interface ContactListResponse {
  data: Contact[];
  total: number;
  page: number;
  totalPages: number;
}