# lib/api-zod/src/generated/types/demoBookingListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Finventory API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DemoBooking } from './demoBooking';

export interface DemoBookingListResponse {
  data: DemoBooking[];
  total: number;
  page: number;
  totalPages: number;
}